requirements.txt
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
imbalanced-learn==0.14.1
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.9
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pillow==12.2.0
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scapy==2.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.5
threadpoolctl==3.6.0
tqdm==4.67.3
